requirements.txt
flask
pyjwt>=2.0
bcrypt
python-dotenv
